requirements.txt
absl-py==0.13.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==12.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
blosc==1.11.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==3.0.2
debugpy==1.8.13
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.11.1
dill==0.3.5.1
dm-control==1.0.8
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.4.1
entrypoints==0.4
etils==1.3.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.30.2
hydra-core==1.2.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-packaging==0.11.1
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab==3.0.14
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
labmaze==1.0.6
lazy_loader==0.4
llvmlite==0.39.1
loguru==0.7.3
lxml==5.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mistune==3.1.3
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.2.3
multidict==6.1.0
narwhals==1.34.0
natsort==8.4.0
nbclassic==0.5.6
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.56.4
numcodecs==0.12.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.5.3.56
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
portalocker==3.0.0
proglog==0.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.2
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch3d==0.3.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.20.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-video==1.1.11
scipy==1.10.1
Send2Trash==1.8.3
sentry-sdk==2.25.1
setproctitle==1.3.5
Shapely==1.8.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.7
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.1
torch-geometric==2.6.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typeguard==4.4.0
typing_extensions==4.13.1
tyro==0.9.18
tzdata==2025.2
urllib3==2.2.3
visdom==0.2.4
wandb==0.19.9
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yacs==0.1.8
yapf==0.31.0
yarl==1.15.2
zarr==2.12.0
zipp==3.20.2
